requirements.txt
backports.tarfile==1.2.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-speech==2.26.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
gTTS==2.5.1
httplib2==0.22.0
idna==3.7
importlib_metadata==8.4.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.8.0
keyring==25.3.0
more-itertools==10.4.0
oauthlib==3.2.2
playwright==1.46.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.20.0
pyee==11.1.0
pygame==2.6.0
pyparsing==3.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
SecretStorage==3.3.3
SpeechRecognition==3.10.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
zipp==3.20.1
